Add StyleFactory and TableFactory for consistent UI styling and table creation
- Implemented StyleFactory class to define consistent styles for tabs, buttons, input fields, tables, and dialogs.
- Created TableFactory class to facilitate the creation and configuration of QTableWidget instances with common settings.
- Integrated styles from StyleFactory into TableFactory for a cohesive look and feel across the application.
