+++ app/ui/pages/inventory/__init__.py
@@ -1,0 +1,1 @@
+from .inventory_page import InventoryPage
