+++ app/ui/pages/inventory/dialogs/__init__.py
@@ -1,0 +1,3 @@
+from .base_dialog import BaseDialog
+from .product_dialog import ProductDialog
+from .service_dialog import ServiceDialog
